+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * Added variable-name subsetting to `[.tableby()`. One can now subset by logicals, numerics, or character vectors.
 
+* Fixed a bug in `tableby()` related to zero-length factor levels. (#22)
+
 # arsenal 0.5.0
 
 * Code all now in GitHub; issues can now be submitted there. Checking is now performed automatically on Travis-CI.